Fix casting message length to unsigned int
Converting a `byte` with `toInt()` gives negative numbers above 127.

First, we need to convert the signed `byte` to an unsigned byte and
then we can cast it to an `int` to get the true native byte value.

+++ src/main/kotlin/com/kurzdigital/vds/vds/Decoder.kt
@@ -36,7 +36,7 @@ private fun ByteBuffer.readVds(): Vds {
     while (true) {
         val pos = position()
         val tag = get()
-        val value = getByteArray(get().toInt())
+        val value = getByteArray(get().toUByte().toInt())
         if (tag == 0xff.toByte()) {
             rewind()
             val (id, mapper) = header.getMapper()
